provision: move random.so into standalone rpm
To avoid pulling in build dependencies in minimal testing environment.
