+++ .copr/Makefile
@@ -1,0 +1,3 @@
+srpm:
+	make -C ./src/rpms srpm
+	[ -n "$(outdir)" ] && cp ./rpmbuild/SRPMS/* "$(outdir)" || :
